+++ lib/weapon-normalizer.test.ts
