+++ lib/weapon-normalizer.ts
